fix: Hero section image (photos) and title in Features section
