+++ components/drive/HeroSection.tsx
@@ -6,7 +6,7 @@ import Header from '../shared/Header';
 
 const HeroSection = ({ textContent, lang, download }) => (
   <section className="flex w-full flex-col">
-    <div className="flex flex-col items-center py-40">
+    <div className="flex flex-col items-center pb-32 pt-40">
       {/* Main title */}
       <div className="flex flex-col items-center justify-center space-y-6 px-5 text-center">
         <div className="flex w-max items-center justify-center rounded-lg bg-gray-5 py-2 px-4">
