+++ components/photos/FeaturesSection.tsx
@@ -14,7 +14,7 @@ const FeaturesSection = ({ textContent, lang, device, download }) => (
       <Parallax
         className="flex w-full flex-col items-center justify-center text-center"
         renderLayer={(percentage) => (
-          <h2 className="relative mb-10 text-3xl font-medium text-cool-gray-90 lg:mb-20 lg:text-5xl">
+          <h2 className="relative mb-10 text-3xl font-semibold text-cool-gray-90 lg:mb-20 lg:text-5xl">
             <Transition
               show={anim.trigger(percentage)}
               enter="transition-all duration-500"
